Split PageItemEditor into logical and layout components, and added story
